MessageBatch -> BasicPublishBatch
BasicPublishBatch can be used for more efficient batch publishing that
will better utilise TCP and provides more throughput.

[#152041636]
