+++ projects/client/Unit/src/unit/TestBasicPublishBatch.cs
@@ -1,4 +1,4 @@
-// This source code is dual-licensed under the Apache License, version
+﻿// This source code is dual-licensed under the Apache License, version
 // 2.0, and the Mozilla Public License, version 1.1.
 //
 // The APL v2.0:
@@ -37,11 +37,29 @@
 //  The Initial Developer of the Original Code is Pivotal Software, Inc.
 //  Copyright (c) 2007-2016 Pivotal Software, Inc.  All rights reserved.
 //---------------------------------------------------------------------------
-namespace RabbitMQ.Client
+
+using NUnit.Framework;
+using RabbitMQ.Client;
+using RabbitMQ.Client.Impl;
+using System;
+
+namespace RabbitMQ.Client.Unit
 {
-    public interface IMessageBatch
+    internal class TestBasicPublishBatch : IntegrationFixture
     {
-        void Add(string exchange, string routingKey, bool mandatory, IBasicProperties properties, byte[] body);
-        void Publish();
+        [Test]
+        public void TestBasicPublishBatchSend()
+        {
+            Model.QueueDeclare(queue: "test-message-batch-a", durable: false);
+            Model.QueueDeclare(queue: "test-message-batch-b", durable: false);
+            var batch = Model.CreateBasicPublishBatch();
+            batch.Add("", "test-message-batch-a", false, null, new byte [] {});
+            batch.Add("", "test-message-batch-b", false, null, new byte [] {});
+            batch.Publish();
+            var resultA = Model.BasicGet("test-message-batch-a", true);
+            Assert.NotNull(resultA);
+            var resultB = Model.BasicGet("test-message-batch-b", true);
+            Assert.NotNull(resultB);
+        }
     }
 }
